+++ R/TRI0-model.R
@@ -7,27 +7,35 @@
 TRI0 <- function(x, ...) {
   UseMethod("TRI0")
 }
-
+#' @name TRI0
+#' @export
+TRI0.PATH0 <- function(x, ...) {
+ TRI0(PATH(x), ...)
+}
+#' @name TRI0
+#' @export
 TRI0.PATH <- function(x, ...) {
   vertex <- x$vertex
   if (nrow(vertex) < 3) stop("need at least 3 coordinates")
   if (anyNA(vertex$x_)) stop("missing values in x_")
   if (anyNA(vertex$y_)) stop("missing values in y_")
-  if (all(x$path$ncoords_ < 2)) stop("TRI for PATH cannot include degenerate paths, see '.$path$ncoords_'")
-  if (any(x$path$ncoords_ < 3)) {
+  path <- sc_path(x)
+  if (all(path$ncoords_ < 2)) stop("TRI for PATH cannot include degenerate paths, see '.$path$ncoords_'")
+  if (any(path$ncoords_ < 3)) {
     warning("filtering out paths with fewer than 3 coordinates before attempting triangulation by ear clipping")
-    x <- x$path %>% dplyr::filter(.data$ncoords_ > 2)
+    path <- path %>% dplyr::filter(.data$ncoords_ > 2)
   }
   ## pretty sure I'll live to regret this ...
   ## (but the right alternative is a smart DEL visibility classifier )
   ## if we get lines, just pretend they all independently POLYGON
-  if (!"subobject" %in% names(x$path)) {
+  if (!"subobject" %in% names(path)) {
     warning("assuming that all paths are independent (i.e. all islands, no holes)")
     ##x$path$subobject <- 1
-    x$path <- x$path %>% dplyr::group_by(.data$object_) %>%
-      dplyr::mutate(subobject = row_number()) %>%
+    path <- path %>% dplyr::group_by(.data$object_) %>%
+      dplyr::mutate(subobject = dplyr::row_number()) %>%
       dplyr::ungroup()
   }
+  x$path <- path
   tri <- triangulate_0(x)
   tri$visible <- TRUE
 
@@ -48,13 +56,13 @@ TRI0.PATH <- function(x, ...) {
 triangulate_0 <- function(x, ...) {
   objlist <- split(x$path, x$path$object_)
   objlist <- objlist[unique(x$path$object_)]
-  polygon_count <- nrow(dplyr::distinct(x$path[c("object", "subobject")]))
+  polygon_count <- nrow(dplyr::distinct(x$path[c("object_", "subobject_")]))
   trilist <- vector("list", polygon_count)
   itri <- 0
   for (i in seq_along(objlist)) {
     obj <- objlist[[i]]
-    subobjlist <- split(obj, obj$subobject)
-    subobjlist <- subobjlist[unique(obj$subobject)]
+    subobjlist <- split(obj, obj$subobject_)
+    subobjlist <- subobjlist[unique(obj$subobject_)]
     for (j in seq_along(subobjlist)) {
       itri <- itri + 1
       verts <- subobjlist[[j]] %>%