+++ NEWS.md
@@ -1,8 +1,10 @@
 # silicate 0.1.1.9001
 
+* Models now have print methods. 
+
 * The  `filter` method for SC is now exported. 
 
-* Incorporate new `BINARY` model, a better starting point for decomposing objects. 
+* Incorporate new version of `SC0` model, a better starting point for decomposing objects. 
 
 * SC can now interpret TRI, which allows for very easy DEL re-triangulations and fine
  control over the quality and details of the mesh